[FEAT] 에러 처리

+++ src/main/java/drive_only/drive_only_server/service/comment/CommentService.java
@@ -11,6 +11,10 @@ import drive_only.drive_only_server.dto.comment.update.CommentUpdateRequest;
 import drive_only.drive_only_server.dto.comment.update.CommentUpdateResponse;
 import drive_only.drive_only_server.dto.common.PaginatedResponse;
 import drive_only.drive_only_server.dto.meta.Meta;
+import drive_only.drive_only_server.exception.custom.CommentNotFoundException;
+import drive_only.drive_only_server.exception.custom.CourseNotFoundException;
+import drive_only.drive_only_server.exception.custom.OwnerMismatchException;
+import drive_only.drive_only_server.exception.custom.ParentCommentNotFoundException;
 import drive_only.drive_only_server.repository.comment.CommentRepository;
 import drive_only.drive_only_server.repository.course.CourseRepository;
 import drive_only.drive_only_server.security.LoginMemberProvider;
@@ -66,6 +70,7 @@ public class CommentService {
     @Transactional
     public CommentUpdateResponse updateComment(Long commentId, CommentUpdateRequest request) {
         Comment comment = findComment(commentId);
+        validateCommentOwner(comment);
         comment.update(request);
         return new CommentUpdateResponse(comment.getId(), SUCCESS_UPDATE);
     }
@@ -73,23 +78,28 @@ public class CommentService {
     @Transactional
     public CommentDeleteResponse deleteComment(Long commentId) {
         Comment comment = findComment(commentId);
+        validateCommentOwner(comment);
         comment.clearChildComments();
         commentRepository.delete(comment);
         return new CommentDeleteResponse(comment.getId(), SUCCESS_DELETE);
     }
 
     private Course findCourse(Long courseId) {
-        return courseRepository.findById(courseId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 코스(게시글)을 찾을 수 없습니다."));
+        return courseRepository.findById(courseId).orElseThrow(CourseNotFoundException::new);
     }
 
     private Comment findComment(Long commentId) {
-        return commentRepository.findById(commentId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 댓글을 찾을 수 없습니다."));
+        return commentRepository.findById(commentId).orElseThrow(CommentNotFoundException::new);
     }
 
     private Comment findParentComment(Long parentCommentId) {
-        return commentRepository.findById(parentCommentId)
-                .orElseThrow(() -> new IllegalArgumentException("부모 댓글을 찾을 수 없습니다."));
+        return commentRepository.findById(parentCommentId).orElseThrow(ParentCommentNotFoundException::new);
+    }
+
+    private void validateCommentOwner(Comment comment) {
+        Member loginMember = loginMemberProvider.getLoginMember();
+        if (!comment.isWrittenBy(loginMember)) {
+            throw new OwnerMismatchException();
+        }
     }
 }
